Implemented GFGeometryCollection isValid algorithm. - Converted isValidOperation into a first class function call instead of a visitor. - Implemented GFGeometryCollection isValid algorithm. - Added GFGeometryCollection isValid tests.
