+++ GeoFeatures/GFGeometry.mm
@@ -36,7 +36,7 @@
 #include "internal/geofeatures/operators/PerimeterOperation.hpp"
 #include "internal/geofeatures/operators/WKTOperation.hpp"
 #include "internal/geofeatures/operators/WithinOperation.hpp"
-#include "internal/geofeatures/operators/IsValidOperation.hpp"
+#include "internal/geofeatures/operators/IsValid.hpp"
 #include "internal/geofeatures/operators/IntersectsSelfOperation.hpp"
 #include "internal/geofeatures/operators/Intersects.hpp"
 
@@ -92,7 +92,7 @@ namespace gf = geofeatures;
         try {
             const auto variant = [self cppGeometryPtrVariant];
 
-            return boost::apply_visitor(gf::operators::IsValidOperation(), variant);
+            return gf::operators::isValid(variant);
 
         } catch (std::exception & e) {
             @throw [NSException exceptionWithName:@"Exception" reason: [NSString stringWithUTF8String: e.what()] userInfo:nil];
